src/polymorphic/contrib/drf/requirements/testing.txt
pytest==3.3.1
pytest-django==3.1.2
pytest-mock==1.6.3
pytest-cov==2.5.1
pytest-sugar==0.9.0
